color formatting complete on results. Results page complete with exception of some minor data manipulation.

+++ StockTradingApp/Data/ViewModels/NetResultViewModel.cs
@@ -4,9 +4,9 @@ namespace StockTradingApp.Data.ViewModels
 {
     public class NetResultViewModel
     {
-        public double Pl_dollar {get; set; }
+        public decimal Pl_dollar {get; set; }
 
-        public double Pl_percent { get; set; }
+        public decimal Pl_percent { get; set; }
 
         public int Num_wins { get; set; }
 
@@ -14,9 +14,9 @@ namespace StockTradingApp.Data.ViewModels
 
         public int Total_trades {  get; set; }
 
-        public double Lg_win { get; set; }
+        public decimal Lg_win { get; set; }
 
-        public double Lg_loss { get; set; }
+        public decimal Lg_loss { get; set; }
 
         public int Win_hold { get; set; }
